finalize contact author email workflow
- add ordering to CodeArchiveStatus enum with respect to which status
overrides another status. For example if a given Publication has
multiple CodeArchiveURLs with ARCHIVED but a single CodeArchiveURL with
NOT_AVAILABLE or NOT_IN_ARCHIVE, we return NOT_AVAILABLE or
NOT_IN_ARCHIVE. Current ordering uses < but perhaps we should switch
that to > to be more intuitive
- add simple aggregate function function that uses the CodeArchiveStatus
ordering to determine if a Publication's code archival status is
NOT_AVAILABLE, NOT_IN_ARCHIVE, or ARCHIVED
- add is_active to CodeArchiveUrl to allow curators to mark a
CodeArchiveUrl as defunct so it won't be included in any calculus about
whether a Publication needs to be updated. This lets us keep it around
for posterity / provenance